Refactor MapContainer
